// src/drivers/mouse-types.ts
export interface MouseStatus {
// ...
  liftOffDistance: "Low" | "Medium" | "High" | null;
// ...
}